refactor: make scroll down arrow a separate component
