+++ src/sections/sponsorship/Sponsorship.Info.tsx
@@ -1,15 +1,8 @@
-import { useState, useEffect } from "react";
 import Image from "next/image";
-import ArrowIcon from "public/assets/arrow.svg";
+
+import ScrollDownArrow from "src/components/ScrollDownArrow";
 
 const SponsorshipInfo: React.FC<{ image: string }> = ({ image }) => {
-  const [scrolled, setScrolled] = useState(false);
-
-  useEffect(() => {
-    const onScroll = () => setScrolled(window.scrollY !== 0);
-    window.addEventListener("scroll", onScroll);
-    return () => window.removeEventListener("scroll", onScroll);
-  }, []);
 
   return (
     <section>
@@ -43,19 +36,7 @@ const SponsorshipInfo: React.FC<{ image: string }> = ({ image }) => {
           <span className="purple"> innovators</span>!
         </h2>
       </div>
-      {scrolled ? null : (
-        <img
-          className="arrow"
-          onClick={() =>
-            window.scrollBy({
-              top: window.innerHeight - 78,
-              behavior: "smooth",
-            })
-          }
-          alt=""
-          src={ArrowIcon.src}
-        />
-      )}
+      <ScrollDownArrow />
     </section>
   );
 };
